test(weave): share BlockingTraceServer + paused via conftest
Move BlockingTraceServer and the paused context manager out of
test_save_nested_defer.py and into tests/trace/conftest.py so other
trace tests can reuse the pause-the-server pattern.
